makefile

+++ LIBFT/ft_strlcat.c
@@ -6,13 +6,13 @@
 /*   By: jaatif-a <jamalch2468@gmail.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/18 18:54:37 by jaatif-a          #+#    #+#             */
-/*   Updated: 2025/09/24 18:15:29 by jaatif-a         ###   ########.fr       */
+/*   Updated: 2025/09/27 19:43:46 by jaatif-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlcat(char *dest, const char *src, size_t size)
+size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
 	size_t	i;
 	size_t	dest_len;
