+++ LIBFT/ft_strlcpy.c
@@ -6,13 +6,13 @@
 /*   By: jaatif-a <jamalch2468@gmail.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/18 18:45:01 by jaatif-a          #+#    #+#             */
-/*   Updated: 2025/09/24 17:20:59 by jaatif-a         ###   ########.fr       */
+/*   Updated: 2025/09/27 19:43:51 by jaatif-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlcpy(char *dest, const char *src, size_t size)
+size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
 	size_t	i;
 	size_t	n;
